+++ resources/js/cp.ts
@@ -17,6 +17,8 @@ import './modules/generated-fields';
 import './modules/component-select';
 import './modules/grouped-entry-type-manager';
 import './modules/queue';
+import './modules/slideout';
+import './modules/auth-method-setup';
 
 window.Cp = Cp as unknown as typeof window.Cp;